rename router.js -> to index.js
